create new branch for resource_management improvements
first steps:
	use allocators in mal interpreter (therefor also the VAL* functions can now optionaly have an allocator as first argument)
	use temp - allocators instead of temp GDKmalloc's in optimizers

+++ sql/scripts/23_skyserver.sql
@@ -35,7 +35,7 @@ external name sql.ms_trunc;
 
 grant execute on function MS_TRUNC to public;
 
-CREATE FUNCTION MS_ROUND(num double, prc int, truncat int)
+CREATE FUNCTION MS_ROUND(num double, prc tinyint, truncat int)
 RETURNS double
 BEGIN
 	IF (truncat = 0)
@@ -46,7 +46,7 @@ END;
 
 grant execute on function MS_ROUND to public;
 
-CREATE FUNCTION MS_STR(num float, prc int, truncat int)
+CREATE FUNCTION MS_STR(num float, prc tinyint, truncat int)
 RETURNS string
 BEGIN
 	RETURN CAST(num as string);
